seperated functions, fixed tests, added tests to check when function called
