+++ packages/gafl-webapp-service/src/pages/licence-details/licence-length/__tests__/update-transaction.spec.js
@@ -1,16 +1,20 @@
 import updateTransaction from '../update-transaction'
 import { LICENCE_FULFILMENT } from '../../../../uri.js'
 import * as concessionHelper from '../../../../processors/concession-helper.js'
-import { assignPermit } from '../../../../processors/find-and-hash-permit.js'
+import { assignPermit } from '../../../../processors/assign-permit.js'
 import { isPhysical } from '../../../../processors/licence-type-display.js'
 import * as mappings from '../../../../processors/mapping-constants.js'
 import moment from 'moment-timezone'
+import { hashPermission } from '../../../../processors/hash-permission.js'
 
 jest.mock('../../../../processors/concession-helper.js')
-jest.mock('../../../../processors/find-and-hash-permit.js', () => ({
+jest.mock('../../../../processors/assign-permit.js', () => ({
   assignPermit: jest.fn((_permission, _request) => {})
 }))
 jest.mock('../../../../processors/licence-type-display.js')
+jest.mock('../../../../processors/hash-permission.js', () => ({
+  hashPermission: jest.fn((_permission, _request) => {})
+}))
 
 describe('licence-details > update-transaction', () => {
   describe('default', () => {
@@ -33,8 +37,27 @@ describe('licence-details > update-transaction', () => {
       }))
     })
 
+    it('calls hashPermission with parameter of permission and request', async () => {
+      const mockPermission = { licenceLength: '12M' }
+      hashPermission.mockReturnValue(mockPermission)
+      assignPermit.mockReturnValue(mockPermission)
+      const request = getMockRequest('12M', mockPermission)
+      await updateTransaction(request)
+      expect(assignPermit).toBeCalledWith(mockPermission, request)
+    })
+
+    it('calls assignPermit with return value of hashPermission and request', async () => {
+      const mockPermission = { licenceLength: '12M' }
+      hashPermission.mockReturnValue(mockPermission)
+      assignPermit.mockReturnValue(mockPermission)
+      const request = getMockRequest('12M', mockPermission)
+      await updateTransaction(request)
+      expect(assignPermit).toBeCalledWith(mockPermission, request)
+    })
+
     it('should find the permit with the available data', async () => {
       const permission = { licenceLength: '1D' }
+      hashPermission.mockReturnValue({ licenceLength: '1D' })
       assignPermit.mockReturnValue({ licenceLength: '12M' })
       const request = getMockRequest('1D', permission)
       await updateTransaction(request)
@@ -43,6 +66,7 @@ describe('licence-details > update-transaction', () => {
     })
 
     it('should set the licence fulfilment page to false on the status', async () => {
+      hashPermission.mockReturnValue({ licenceLength: '12M' })
       assignPermit.mockReturnValue({ licenceLength: '12M' })
       const request = getMockRequest()
       await updateTransaction(request)
@@ -52,16 +76,18 @@ describe('licence-details > update-transaction', () => {
     })
 
     it('should set the transaction to update with the permit', async () => {
-      const mockPermit = Symbol('permit')
-      assignPermit.mockReturnValue(mockPermit)
+      const mockPermission = Symbol('permission')
+      hashPermission.mockReturnValue(mockPermission)
+      assignPermit.mockReturnValue(mockPermission)
       const permission = { licenceLength: '12M' }
       const request = getMockRequest('12M', permission)
       await updateTransaction(request)
-      expect(request.cache.mock.results[3].value.helpers.transaction.setCurrentPermission).toHaveBeenCalledWith(mockPermit)
+      expect(request.cache.mock.results[3].value.helpers.transaction.setCurrentPermission).toHaveBeenCalledWith(mockPermission)
     })
 
     describe('checkLicenceToStart', () => {
       it('should set licenceStartTime to null when the length is 12M', async () => {
+        hashPermission.mockReturnValue({ licenceLength: '12M' })
         assignPermit.mockReturnValue({ licenceLength: '12M' })
         const permission = { licenceLength: '12M' }
         const request = getMockRequest('12M', permission)
@@ -75,6 +101,7 @@ describe('licence-details > update-transaction', () => {
       })
 
       it('should set licenceToStart to after-payment when the length is 12M and the start date matches the current day', async () => {
+        hashPermission.mockReturnValue({ licenceLength: '12M', licenceStartDate: moment() })
         assignPermit.mockReturnValue({ licenceLength: '12M', licenceStartDate: moment() })
         const permission = { licenceLength: '12M', licenceStartDate: moment() }
         const request = getMockRequest('12M', permission)
@@ -89,6 +116,7 @@ describe('licence-details > update-transaction', () => {
 
       it('should not modify licenceToStart when the length is 12M and the start date is a later day', async () => {
         const licenceToStart = Symbol('licenceToStart')
+        hashPermission.mockReturnValue({ licenceLength: '1D', licenceStartDate: '2100-01-01', licenceToStart })
         assignPermit.mockReturnValue({ licenceLength: '1D', licenceStartDate: '2100-01-01', licenceToStart })
         const permission = { licenceLength: '12M', licenceStartDate: '2100-01-01', licenceToStart }
         const request = getMockRequest('12M', permission)
@@ -103,6 +131,7 @@ describe('licence-details > update-transaction', () => {
 
       it('should not modify licenceStartTime when the length is not 12M', async () => {
         const licenceStartTime = Symbol('licenceStartTime')
+        hashPermission.mockReturnValue({ licenceLength: '1D', licenceStartTime })
         assignPermit.mockReturnValue({ licenceLength: '1D', licenceStartTime })
         const permission = { licenceLength: '1D', licenceStartTime }
         const request = getMockRequest('1D', permission)
@@ -119,6 +148,7 @@ describe('licence-details > update-transaction', () => {
     describe('checkContactDetails', () => {
       it('should set the correct values when the permit is physical and preferredMethodOfConfirmation and/or preferredMethodOfReminder are set to none', async () => {
         isPhysical.mockReturnValueOnce(true)
+        hashPermission.mockReturnValue({ licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.none }, licenceLength: '12M' })
         assignPermit.mockReturnValue({ licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.none }, licenceLength: '12M' })
         const permission = { licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.none } }
         const request = getMockRequest('12M', permission)
@@ -137,6 +167,7 @@ describe('licence-details > update-transaction', () => {
 
       it('should set the correct values the permit is not physical and preferredMethodOfConfirmation and/or preferredMethodOfReminder are set to letter', async () => {
         isPhysical.mockReturnValueOnce(false)
+        hashPermission.mockReturnValue({ licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.letter }, licenceLength: '12M' })
         assignPermit.mockReturnValue({ licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.letter }, licenceLength: '12M' })
         const permission = { licensee: { preferredMethodOfConfirmation: mappings.HOW_CONTACTED.letter } }
         const request = getMockRequest('12M', permission)
@@ -158,6 +189,7 @@ describe('licence-details > update-transaction', () => {
       it('should store the concession to previouslyDisabled when the licenceLength is not 12M and the permission has a disabled concession', async () => {
         concessionHelper.hasDisabled.mockReturnValueOnce(true)
         const concession = { type: mappings.CONCESSION.DISABLED }
+        hashPermission.mockReturnValue({ licenceLength: '1D', concessions: [concession] })
         assignPermit.mockReturnValue({ licenceLength: '1D', concessions: [concession] })
         const permission = { licenceLength: '1D', concessions: [concession] }
         const request = getMockRequest('1D', permission)
@@ -173,18 +205,20 @@ describe('licence-details > update-transaction', () => {
       it('should call removeDisabled when the licenceLength is not 12M and the permission has a disabled concession', async () => {
         concessionHelper.hasDisabled.mockReturnValueOnce(true)
         const concession = { type: mappings.CONCESSION.DISABLED }
-        const mockPermit = { licenceLength: '1D', concessions: [concession] }
-        assignPermit.mockReturnValue(mockPermit)
+        const mockPermission = { licenceLength: '1D', concessions: [concession] }
+        hashPermission.mockReturnValue(mockPermission)
+        assignPermit.mockReturnValue(mockPermission)
         const permission = { licenceLength: '1D', concessions: [concession] }
         const request = getMockRequest('1D', permission)
         await updateTransaction(request)
 
-        expect(concessionHelper.removeDisabled).toHaveBeenCalledWith(mockPermit)
+        expect(concessionHelper.removeDisabled).toHaveBeenCalledWith(mockPermission)
       })
 
       it('should re-add the disabled concession and clear previouslyDisabled when the licenceLength is 12M and the permission previously had a disabled concession', async () => {
         concessionHelper.hasDisabled.mockReturnValueOnce(false)
         const previousConcession = Symbol('previousConcession')
+        hashPermission.mockReturnValue({ licenceLength: '12M', concessions: [], previouslyDisabled: previousConcession })
         assignPermit.mockReturnValue({ licenceLength: '12M', concessions: [], previouslyDisabled: previousConcession })
         const permission = { licenceLength: '12M', concessions: [], previouslyDisabled: previousConcession }
         const request = getMockRequest('12M', permission)
@@ -199,6 +233,7 @@ describe('licence-details > update-transaction', () => {
       })
 
       it('should set the number of rods to 2 when the licenceLength is not 12M and the licenceType is trout and course', async () => {
+        hashPermission.mockReturnValue({ licenceLength: '1D', licenceType: 'Trout and coarse' })
         assignPermit.mockReturnValue({ licenceLength: '1D', licenceType: 'Trout and coarse' })
         const permission = { licenceLength: '1D', licenceType: mappings.LICENCE_TYPE['trout-and-coarse'], numberOfRods: '10' }
         const request = getMockRequest('1D', permission)
