+++ packages/gafl-webapp-service/src/processors/hash-permission.js
@@ -1,8 +1,7 @@
-import filterPermits from './filter-permits.js'
 import crypto from 'crypto'
 import db from 'debug'
 
-const debug = db('webapp:find-and-hash-permit')
+const debug = db('webapp:hash-permission')
 
 export const hashPermission = async (permission, request) => {
   const clonedPermission = { ...permission }
@@ -22,20 +21,4 @@ export const hashPermission = async (permission, request) => {
 
   await request.cache().helpers.transaction.setCurrentPermission(clonedPermission)
   return clonedPermission
-}
-
-export const assignPermit = async (permission, request) => {
-  const hashedPermission = await hashPermission(permission, request)
-  const permit = await filterPermits(permission)
-  hashedPermission.permit = permit
-  if (permit) {
-    if (!permit.newCostStartDate || !permit.newCost) {
-      debug('permit missing new cost details', permission)
-    }
-  } else {
-    debug("permit wasn't retrieved", permission)
-  }
-
-  await request.cache().helpers.transaction.setCurrentPermission(hashedPermission)
-  return hashedPermission
 }
